pyproject.toml
[tool.poetry]
name = "word2vec-demo"
version = "0.1.0"
description = "this is word2vec demo script"
authors = ["lovelovetrb"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
gensim = "^4.3.1"
sudachipy = "^0.6.7"
ginza = "^5.1.2"
ja-ginza = "^5.1.2"
sklearn = "^0.0.post5"
pandas = "^2.0.1"
wikiextractor = "^3.0.6"
mecab-python3 = "^1.0.6"
pyyaml = "^6.0"
unidic-lite = "^1.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
